requirements.txt
twitchio==1.1.0
PyYAML==5.3.1
